Score guesses according to how evenly they divide up the solution space
Each character of a guess can be exactly right (🟩), in the wrong position (🟨), or not in the word at all (⬛). Thus, there are 243 = 3^5 possible results/buckets.

Here we test each (remaining) word in the list against every other word, and look at the distribution of the results into these 243 buckets.

So first we calculate the mean (average) number of results in a bucket, and from that the variance. A smaller variance means that the distribution is more even.
Note: Normally the standard deviation would be used for that, which is the square root of the variance. Since we only need an ordering the variance is good enough.

This is heavily inspired by Sean Plays https://youtu.be/BN-Yan03m8s