Simplified game state by removing tracking of chars
Tracking individual chars is error-prone. For example, guessing
'feued' with secret 'piper' yields hint '⬛⬛⬛🟩⬛'. The first 'e' was
not properly marked as illegal on all but the second to last position.

It's easier to keep track of the set of remaining viable solutions.
Solution sets are already cached by guess and hint, and each new
guess and hint pair yield a new set of solutions, which needs to be
intersected with the remaining viable solutions.