Re-tested previously used best-word methods
expected_remaining_solution_counts(…) is just as stable as (and
basically identical to) fewest_remaining_solutions_for_game(…) if
the division by the total number of solutions is done only once, over
the sum as usize.

No surprise really, as sums of floating point numbers are prone to
rounding errors, depending on the order of terms – it's best to avoid.