Fully replaced hashmap based caches, and fixed a bug
When there are only one or two solutions left in a subtree, one must
check if the guess is part of the remaining solutions. If so, the guess
might have been correct this turn already, rather than on the next turn.
Because of this, all calculations that used count_by_attempts(…),
turn_sums(…) or fewest_remaining_solutions(…) were slightly wrong
(reported too high of a score, where lower is better)